Bring `breadboard/remote` types to follow the now-common pattern.
