+++ packages/breadboard/src/remote/index.ts
@@ -12,9 +12,8 @@ export {
 } from "./worker.js";
 export { ProxyServer, ProxyClient } from "./proxy.js";
 export { RunServer, RunClient } from "./run.js";
-export type { ClientRunResult } from "./run.js";
 export { InitServer, InitClient } from "./init.js";
 export { defineConfig, hasOrigin, type ProxyServerConfig } from "./config.js";
-export type * from "./protocol.js";
+export type * from "./types.js";
 export type * from "./config.js";
 export type * from "./http.js";
