+++ packages/breadboard/src/remote/run.ts
@@ -21,14 +21,14 @@ import {
   RunState,
 } from "../types.js";
 import {
-  AnyProbeMessage,
+  AnyClientRunResult,
   AnyRunRequestMessage,
   AnyRunResponseMessage,
-  ClientTransport,
   InputResolveRequest,
+  RunClientTransport,
   RunRequestMessage,
   ServerTransport,
-} from "./protocol.js";
+} from "./types.js";
 
 const resumeRun = (request: AnyRunRequestMessage) => {
   const [type, , state] = request;
@@ -127,35 +127,6 @@ export class RunServer {
     }
   }
 }
-
-type RunClientTransport = ClientTransport<
-  AnyRunRequestMessage,
-  AnyRunResponseMessage
->;
-
-type ReplyFunction = {
-  reply: (chunk: AnyRunRequestMessage[1]) => Promise<void>;
-};
-
-type ClientRunResultFromMessage<ResponseMessage> = ResponseMessage extends [
-  string,
-  object,
-  RunState?,
-]
-  ? {
-      type: ResponseMessage[0];
-      data: ResponseMessage[1];
-      state?: RunState;
-    } & ReplyFunction
-  : never;
-
-export type AnyClientRunResult =
-  ClientRunResultFromMessage<AnyRunResponseMessage>;
-
-export type AnyProbeClientRunResult =
-  ClientRunResultFromMessage<AnyProbeMessage>;
-
-export type ClientRunResult<T> = T & ReplyFunction;
 
 const createRunResult = (
   response: WritableResult<AnyRunResponseMessage, AnyRunRequestMessage>
